requirements.txt
# Core libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
numpy>=1.24.0

# Web & UI
streamlit>=1.25.0
pyngrok>=7.0.0

# Utils
python-dotenv>=1.0.0

# Optional but useful
tqdm>=4.65.0
